feature: DeepEqualNodes
DeepEqualNodes allows two slices of nodes to be compared.

The slices must have the same length (including zero) or the result will always be false. If both slices contain zero elements then the result is always true.

Every node in the left must DeepEqual a node on the right. The same node cannot be used twice in a comparison. The slices are allowed to have duplicate nodes (by reference or value) as long as the other side has an equal amount of duplicates.
